1 - expose a method to register a sub config
