+++ src/taipy/config/__init__.py
@@ -12,9 +12,10 @@
 from .checker.issue import Issue
 from .checker.issue_collector import IssueCollector
 from .config import Config
+from .abstractSubConfig import AbstractSubConfig
+
 from .data_node.data_node_config import DataNodeConfig
 from .global_app.global_app_config import GlobalAppConfig
-from .auth.auth_config import AuthConfig
 from .job_execution.job_config import JobConfig
 from .pipeline.pipeline_config import PipelineConfig
 from .scenario.scenario_config import ScenarioConfig
