+++ src/taipy/config/_config.py
@@ -12,7 +12,6 @@
 from copy import copy
 from typing import Dict
 
-from .auth.auth_config import AuthConfig
 from .data_node.data_node_config import DataNodeConfig
 from .global_app.global_app_config import GlobalAppConfig
 from .job_execution.job_config import JobConfig
@@ -24,20 +23,26 @@ from .task.task_config import TaskConfig
 class _Config:
     DEFAULT_KEY = "default"
 
+    _default_sub_configs = None
+
     def __init__(self):
         self._global_config: GlobalAppConfig = GlobalAppConfig()
-        self._auth_config: AuthConfig = AuthConfig()
+        self._sub_configs = None
         self._job_config: JobConfig = JobConfig()
         self._data_nodes: Dict[str, DataNodeConfig] = {}
         self._tasks: Dict[str, TaskConfig] = {}
         self._pipelines: Dict[str, PipelineConfig] = {}
         self._scenarios: Dict[str, ScenarioConfig] = {}
 
+    def register(self, sub_config):
+        self._sub_configs = sub_config
+        self._default_sub_configs = sub_config.default_config()
+
     @classmethod
     def _default_config(cls):
         config = _Config()
+        config._sub_configs = cls._default_sub_configs
         config._global_config = GlobalAppConfig.default_config()
-        config._auth_config = AuthConfig.default_config()
         config._job_config = JobConfig().default_config()
         config._data_nodes = {cls.DEFAULT_KEY: DataNodeConfig.default_config(cls.DEFAULT_KEY)}
         config._tasks = {cls.DEFAULT_KEY: TaskConfig.default_config(cls.DEFAULT_KEY)}
@@ -46,8 +51,9 @@ class _Config:
         return config
 
     def _update(self, other_config):
+        if self._sub_configs:
+            self._sub_configs._update(other_config._sub_configs._to_dict())
         self._global_config._update(other_config._global_config._to_dict())
-        self._auth_config._update(other_config._auth_config._to_dict())
         self._job_config._update(other_config._job_config._to_dict())
         self.__update_entity_configs(self._data_nodes, other_config._data_nodes, DataNodeConfig)
         self.__update_entity_configs(self._tasks, other_config._tasks, TaskConfig)
